requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
arrow==1.3.0
async-timeout==4.0.3
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.4
boto3==1.34.11
botocore==1.34.11
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
croniter==1.3.15
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.5.14
dateutils==0.6.12
dbus-python==1.2.18
deepdiff==6.7.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.0
fastapi==0.88.0
fasteners==0.14.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gpt-index==0.4.24
gradio==3.48.0
gradio_client==0.6.1
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.3
importlib-metadata==4.6.4
importlib-resources==6.1.1
inquirer==3.1.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.0.148
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.75
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.0.1
lightning-cloud==0.5.57
lightning-utilities==0.10.0
llvmlite==0.41.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
numba==0.58.1
numexpr==2.8.8
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
openai==0.28.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
openapi-schema-pydantic==1.2.4
ordered-set==4.1.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycryptodome==3.19.1
pycups==2.0.1
pydantic==1.10.13
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.1.3
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
readchar==4.0.5
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
SecretStorage==3.3.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
systemd-python==234
tenacity==8.2.3
texttable==1.6.4
tiktoken==0.5.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2
torchmetrics==1.2.1
tqdm==4.66.1
traitlets==5.14.1
triton==2.1.0
ttystatus==0.38
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.25.0
wadllib==1.3.6
wcwidth==0.2.12
websocket-client==1.2.3
websockets==11.0.3
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
